refactor: FaviconService.swift → Services
